Improve QoL when rendering markdown
NOTE: Due to something that I coudn't diagnose, the headers are not
rendered as expected (with the numbers indicating the level).
If I update
`~/.local/share/LazyVim/lazy/LazyVim/lua/lazyvim/plugins/extras/lang/markdown.lua`
to only include the render-markdown plugin, the header is rendered as
expected. Anything else that I keep in the file, the headers break.